Split Hurricane Electric tunnel updates and FreeDNS updates apart.
Now that he.net supports auto-update of a DNS entry from a tunnel update, it's
very useful to have FreeDNS updates be completely optional and independent.

This change also simplifies the code and takes advantage of the saved dns ips
associated with the tunnel, which previously were only written and never read.